Add files via upload
Here are some of the algorithms used in the project. The number on a file refers to the score earned from the MSE. I am not allowed to share the dataframe for confidentiality reasons, but the structure can be a base for any regression problem. In some files you should normalize the input values ​​for best results. ;)